Optimizing topic creation tests: removed waits when marking topic as sticked.
This is done because most of the time we don't have that checkbox present on
the page but Selenium still waits for implicit timeout (10 secs) for it to show
up. Because of that we spend 20 secs just to wait for Announcement & Sticked
checkboxes to appear, but this never happens. Therefore updated the code
that created the topic not to wait for these checkboxes if they are not on the
page RIGHT NOW.
This might sound dangerous since there might be cases when we'd like to check
the correct work of permissions - whether they impact showing of checkboxes
or not, but I think we'd need some other ways to check that - the price of
these tests is too high.