Make @Before & @After methods run always
Because groups were introduced and included in test suites inside of
testng.xml, they @Before/@After methods were not executed. In order
to fix that we should either assign groups to the methods like it was
done for @Test or simply set it as `alwaysRun = true`. Because we need
to run these setup methods all the time, the last option was used.