JC-2108 JC-2109 Delete and Move Topics/Posts created by the same thread.
1. Behavior changed to move and delete only Topics and Posts, created by
the same virtual user (thread) in contrast to taking random ones as it
used to be. This allowed to eliminate an interference between threads
which led to numerous 404 errors when several threads tried to delete
the same post (JC-2109). This new approach is not completely fair, but
it’s similar to the real-life behavior and acceptable for load testing
(at least for tests that launched automatically).
2. The issue JC-2108 (should send a request for whole topic deletion
when need to delete the last post) is also addressed by predictable
deletion of specific (not random) post.
3. The third column (“role”) was added to accounts.txt file to supply
information about an access rights profile of the account. This allows
not to perform actions to which particular user doesn’t have access and
avoid 403 errors. Supported values: admin, moderator, registered. Only
the last one is actually used in the script now.

404 errors are still happen (rarely) when opening random topics, which
may be already deleted by neighbor thread.
All other errors I’ve seen with this new version were real server errors
or their consequences.

It’s a good idea to label last previous version, because this changes
are major and later we should be able to find this point quickly.

Tested on local Maven with both BaseLine (wasn’t changed but use the
same accounts.txt file) and TestPlan.