Changed smallScreen main menu opening check for new iOS
On the isUserLoggedIn() check, somehow, Selenium detects <input
id="searchText"> webElement inside the collapsed main menu as displayed()
in Safari browser and thinks that menu is opened. But when the next check
concerning another existing but collapsed webElement appears <a
id="user-menu-logout"> then this element is not displayed() so the tests
failed with the reason that user is not logged in when expected. This
might be the Safari webdriver issue, so I just managed to change the menu
status check.